+++ src/beat/mod.rs
@@ -26,7 +26,7 @@ use crate::routing::{self, Rule};
 use crate::{
     error::{BeatError, BrokerError, CeleryError},
     protocol::MessageContentType,
-    task::{Signature, Task},
+    task::{Signature, Task, TaskOptions},
 };
 use log::{debug, error, info};
 use std::time::SystemTime;
@@ -56,7 +56,7 @@ where
     broker_connection_retry_delay: u32,
     default_queue: String,
     task_routes: Vec<(String, String)>,
-    content_type: Option<MessageContentType>,
+    task_options: TaskOptions,
 }
 
 /// Used to create a `Beat` app with a custom configuration.
@@ -86,7 +86,7 @@ where
                 broker_connection_retry_delay: 5,
                 default_queue: "celery".into(),
                 task_routes: vec![],
-                content_type: None,
+                task_options: TaskOptions::default(),
             },
             scheduler_backend: LocalSchedulerBackend::new(),
         }
@@ -115,7 +115,7 @@ where
                 broker_connection_retry_delay: 5,
                 default_queue: "celery".into(),
                 task_routes: vec![],
-                content_type: None,
+                task_options: TaskOptions::default(),
             },
             scheduler_backend,
         }
@@ -165,8 +165,8 @@ where
     }
 
     /// Set a default content type of the message body serialization.
-    pub fn content_type(mut self, content_type: MessageContentType) -> Self {
-        self.config.content_type = Some(content_type);
+    pub fn task_content_type(mut self, content_type: MessageContentType) -> Self {
+        self.config.task_options.content_type = Some(content_type);
         self
     }
 
@@ -201,7 +201,7 @@ where
             scheduler_backend: self.scheduler_backend,
             task_routes,
             default_queue: self.config.default_queue,
-            content_type: self.config.content_type,
+            task_options: self.config.task_options,
             broker_connection_timeout: self.config.broker_connection_timeout,
             broker_connection_retry: self.config.broker_connection_retry,
             broker_connection_max_retries: self.config.broker_connection_max_retries,
@@ -221,7 +221,7 @@ pub struct Beat<Br: Broker, Sb: SchedulerBackend> {
     scheduler_backend: Sb,
     task_routes: Vec<Rule>,
     default_queue: String,
-    content_type: Option<MessageContentType>,
+    task_options: TaskOptions,
 
     broker_connection_timeout: u32,
     broker_connection_retry: bool,
@@ -265,23 +265,17 @@ where
     }
 
     /// Schedule the execution of a task.
-    pub fn schedule_task<T, S>(&mut self, signature: Signature<T>, schedule: S)
+    pub fn schedule_task<T, S>(&mut self, mut signature: Signature<T>, schedule: S)
     where
         T: Task + Clone + 'static,
         S: Schedule + 'static,
     {
+        signature.options.update(&self.task_options);
         let queue = match &signature.queue {
             Some(queue) => queue.to_string(),
             None => routing::route(T::NAME, &self.task_routes)
                 .unwrap_or(&self.default_queue)
                 .to_string(),
-        };
-        let signature = if let Some(content_type) = self.content_type {
-            let mut signature = signature;
-            signature.options.content_type = Some(content_type);
-            signature
-        } else {
-            signature
         };
         let message_factory = Box::new(signature);